ml/requirements.txt
streamlit>=1.47.1
fastapi>=0.68.0
uvicorn>=0.15.0
folium>=0.12.0
streamlit-folium>=0.25.0
plotly>=5.0.0
pandas>=1.4.0
numpy>=1.21.0
pillow>=8.3.0
torch>=1.9.0
torchvision>=0.10.0
python-multipart>=0.0.5

# Enhanced Streamlit packages
streamlit-image-comparison>=0.0.4
streamlit-sparrow-labeling>=0.1.1
streamlit-camera-input-live>=0.2.0
streamlit-chatbox>=1.1.13
streamlit-nested-layout>=0.1.4
streamlit-markdown>=1.1.0
streamlit-javascript>=0.1.5
streamlit-feedback>=0.1.4
simplejson>=3.20.1

# Advanced Analytics packages
streamlit-pandas-profiling>=0.1.3
streamlit-searchbox>=0.1.22
st-click-detector>=0.1.3
streamlit-pills>=0.3.0
streamlit-scrollable-textbox>=0.0.3
st-card-component>=0.0.10
pandas-profiling>=3.6.6
ydata-profiling>=4.16.1

# Property Management packages
trubrics>=1.6.2
st-chat-message>=0.3.8
streamlit-awesome-table>=0.1.0
loguru>=0.7.3
typer>=0.16.0
rich>=14.0.0
markdown-it-py>=3.0.0
pygments>=2.19.2
mdurl>=0.1.2
shellingham>=1.5.4

# ML and AI packages for Client Dashboard
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
asyncio-mqtt>=0.16.0
aiohttp>=3.8.0
websockets>=11.0.0
redis>=4.5.0
celery>=5.3.0